+++ src/server-action/create-safe-server-action.ts
@@ -81,7 +81,6 @@ export function createSafeServerAction<
       return {
         message: err.message,
         name: err.name,
-        stack: err.stack,
       }
     }